% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\docType{package}
\name{hctrial}
\alias{hctrial}
\alias{hctrial-package}
\title{hctrial: A package for designing phase 2 clinical trials adjusting for heterogeneous populations.}
\description{
The hctrial package provides functions for designing phase 2 clinical trials that adjust for the
heterogeneity in the population.
}
\details{
Two different ways are considered for designing a trial: based on known subgroups or based on historical data.

For initializing a stratified trial, use \code{strat_start}.

At interim, \code{strat_interim} should be used to adjust the trial.

At the end of the study, \code{strat_end} is used to adjust the trial again.

\code{hist_start}, \code{hist_interim} and \code{hist_end} work analogously, but are based on historical controls.
}
