% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sail.R
\name{sail}
\alias{sail}
\title{Create a source abstraction interface layer (SAIL) object.}
\usage{
sail(path = getwd(), extension = "R", all.files = FALSE,
  use.cache = FALSE, simplify = TRUE, suppressMessages = FALSE)
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path to work in, will be created if it does not exists.}

\item{extension}{[\code{character(1)}]\cr
File extension to work with.
Default is \dQuote{R}.}

\item{all.files}{[\code{logical(1)}]\cr
Also include hidden files, i.e. files whose name start with a dot (\dQuote{.}).
Default is \code{FALSE}.}

\item{use.cache}{[\code{logical(1)}]\cr
Use a memory cache per global default.
Global option which can locally be overwritten in most functions.
Default is \code{FALSE}}

\item{simplify}{[\code{character(1)}]\cr
If only one object is defined in a sourced R file,
should the return value be simplified? If set to \code{TRUE},
instead of a list containing one element the object itself will be returned.}

\item{suppressMessages}{[\code{logical(1)}]\cr
Wrap the \code{\link[base]{sys.source}} command into \code{\link[base]{suppressMessages}}
and \code{link[base]{suppressPackageStartupMessages}}?
Default is \code{FALSE}, i.e. you will see regular output of sourced scripts.}
}
\value{
Object of class \code{sail}. See the documentation of \code{\link{fail}}
  for details.
}
\description{
This function returns an object of class \code{sail} which behaves
like \code{\link{fail}}, but is indented for loading and saving
R source code files.
}

